refactor: refactor web log feature from base scheduler to a new module
