+++ src/memos/mem_scheduler/base_scheduler.py
@@ -25,7 +25,6 @@ from memos.mem_cube.base import BaseMemCube
 from memos.mem_cube.general import GeneralMemCube
 from memos.mem_feedback.simple_feedback import SimpleMemFeedback
 from memos.mem_scheduler.general_modules.init_components_for_scheduler import init_components
-from memos.mem_scheduler.general_modules.misc import AutoDroppingQueue as Queue
 from memos.mem_scheduler.general_modules.scheduler_logger import SchedulerLoggerModule
 from memos.mem_scheduler.memory_manage_modules.retriever import SchedulerRetriever
 from memos.mem_scheduler.monitors.dispatcher_monitor import SchedulerDispatcherMonitor
@@ -37,7 +36,6 @@ from memos.mem_scheduler.schemas.general_schemas import (
     DEFAULT_CONSUME_INTERVAL_SECONDS,
     DEFAULT_CONTEXT_WINDOW_SIZE,
     DEFAULT_MAX_INTERNAL_MESSAGE_QUEUE_SIZE,
-    DEFAULT_MAX_WEB_LOG_QUEUE_SIZE,
     DEFAULT_STARTUP_MODE,
     DEFAULT_THREAD_POOL_MAX_WORKERS,
     DEFAULT_TOP_K,
@@ -46,17 +44,10 @@ from memos.mem_scheduler.schemas.general_schemas import (
     TreeTextMemory_SEARCH_METHOD,
 )
 from memos.mem_scheduler.schemas.message_schemas import (
-    ScheduleLogForWebItem,
     ScheduleMessageItem,
 )
 from memos.mem_scheduler.schemas.monitor_schemas import MemoryMonitorItem
 from memos.mem_scheduler.schemas.task_schemas import (
-    ADD_TASK_LABEL,
-    ANSWER_TASK_LABEL,
-    MEM_ARCHIVE_TASK_LABEL,
-    MEM_ORGANIZE_TASK_LABEL,
-    MEM_UPDATE_TASK_LABEL,
-    QUERY_TASK_LABEL,
     TaskPriorityLevel,
 )
 from memos.mem_scheduler.task_schedule_modules.dispatcher import SchedulerDispatcher
@@ -72,6 +63,7 @@ from memos.mem_scheduler.utils.monitor_event_utils import emit_monitor_event, to
 from memos.mem_scheduler.utils.status_tracker import TaskStatusTracker
 from memos.mem_scheduler.webservice_modules.rabbitmq_service import RabbitMQSchedulerModule
 from memos.mem_scheduler.webservice_modules.redis_service import RedisSchedulerModule
+from memos.mem_scheduler.webservice_modules.web_log_service import WebLogSchedulerModule
 from memos.memories.activation.kv import KVCacheMemory
 from memos.memories.activation.vllmkv import VLLMKVCacheItem, VLLMKVCacheMemory
 from memos.memories.textual.naive import NaiveTextMemory
@@ -93,7 +85,9 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 
-class BaseScheduler(RabbitMQSchedulerModule, RedisSchedulerModule, SchedulerLoggerModule):
+class BaseScheduler(
+    RabbitMQSchedulerModule, RedisSchedulerModule, SchedulerLoggerModule, WebLogSchedulerModule
+):
     """Base class for all mem_scheduler."""
 
     def __init__(self, config: BaseSchedulerConfig):
@@ -122,12 +116,8 @@ class BaseScheduler(RabbitMQSchedulerModule, RedisSchedulerModule, SchedulerLogg
         # optional configs
         self.disabled_handlers: list | None = self.config.get("disabled_handlers", None)
 
-        self.max_web_log_queue_size = self.config.get(
-            "max_web_log_queue_size", DEFAULT_MAX_WEB_LOG_QUEUE_SIZE
-        )
-        self._web_log_message_queue: Queue[ScheduleLogForWebItem] = Queue(
-            maxsize=self.max_web_log_queue_size
-        )
+        self.init_web_log_module(config=self.config)
+
         self._consumer_thread = None  # Reference to our consumer thread/process
         self._consumer_process = None  # Reference to our consumer process
         self._running = False
@@ -855,103 +845,6 @@ class BaseScheduler(RabbitMQSchedulerModule, RedisSchedulerModule, SchedulerLogg
         # Enqueue lower-priority tasks
         if queued_msgs:
             self.memos_message_queue.submit_messages(messages=queued_msgs)
-
-    def _submit_web_logs(
-        self,
-        messages: ScheduleLogForWebItem | list[ScheduleLogForWebItem],
-        additional_log_info: str | None = None,
-    ) -> None:
-        """Submit log messages to the web log queue and optionally to RabbitMQ.
-
-        Args:
-            messages: Single log message or list of log messages
-        """
-        if isinstance(messages, ScheduleLogForWebItem):
-            messages = [messages]  # transform single message to list
-
-        for message in messages:
-            if self.rabbitmq_config is None:
-                return
-            try:
-                # Always call publish; the publisher now caches when offline and flushes after reconnect
-                logger.info(
-                    f"[DIAGNOSTIC] base_scheduler._submit_web_logs: enqueue publish {message.model_dump_json(indent=2)}"
-                )
-                self.rabbitmq_publish_message(message=message.to_dict())
-                logger.info(
-                    "[DIAGNOSTIC] base_scheduler._submit_web_logs: publish dispatched "
-                    "item_id=%s task_id=%s label=%s",
-                    message.item_id,
-                    message.task_id,
-                    message.label,
-                )
-            except Exception as e:
-                logger.error(
-                    f"[DIAGNOSTIC] base_scheduler._submit_web_logs failed: {e}", exc_info=True
-                )
-
-        logger.debug(
-            f"{len(messages)} submitted. {self._web_log_message_queue.qsize()} in queue. additional_log_info: {additional_log_info}"
-        )
-
-    def get_web_log_messages(self) -> list[dict]:
-        """
-        Retrieve structured log messages from the queue and return JSON-serializable dicts.
-        """
-        raw_items: list[ScheduleLogForWebItem] = []
-        while True:
-            try:
-                raw_items.append(self._web_log_message_queue.get_nowait())
-            except Exception:
-                break
-
-        def _map_label(label: str) -> str:
-            mapping = {
-                QUERY_TASK_LABEL: "addMessage",
-                ANSWER_TASK_LABEL: "addMessage",
-                ADD_TASK_LABEL: "addMemory",
-                MEM_UPDATE_TASK_LABEL: "updateMemory",
-                MEM_ORGANIZE_TASK_LABEL: "mergeMemory",
-                MEM_ARCHIVE_TASK_LABEL: "archiveMemory",
-            }
-            return mapping.get(label, label)
-
-        def _normalize_item(item: ScheduleLogForWebItem) -> dict:
-            data = item.to_dict()
-            data["label"] = _map_label(data.get("label"))
-            memcube_content = getattr(item, "memcube_log_content", None) or []
-            metadata = getattr(item, "metadata", None) or []
-
-            memcube_name = getattr(item, "memcube_name", None)
-            if not memcube_name and hasattr(self, "_map_memcube_name"):
-                memcube_name = self._map_memcube_name(item.mem_cube_id)
-            data["memcube_name"] = memcube_name
-
-            memory_len = getattr(item, "memory_len", None)
-            if memory_len is None:
-                if data["label"] == "mergeMemory":
-                    memory_len = len([c for c in memcube_content if c.get("type") != "postMerge"])
-                elif memcube_content:
-                    memory_len = len(memcube_content)
-                else:
-                    memory_len = 1 if item.log_content else 0
-
-            data["memcube_log_content"] = memcube_content
-            data["memory_len"] = memory_len
-
-            def _with_memory_time(meta: dict) -> dict:
-                enriched = dict(meta)
-                if "memory_time" not in enriched:
-                    enriched["memory_time"] = enriched.get("updated_at") or enriched.get(
-                        "update_at"
-                    )
-                return enriched
-
-            data["metadata"] = [_with_memory_time(m) for m in metadata]
-            data["log_title"] = ""
-            return data
-
-        return [_normalize_item(it) for it in raw_items]
 
     def _message_consumer(self) -> None:
         """
